Moved email uniquness validation to function
