+++ app/serializers/users.js
@@ -1,5 +1,5 @@
 const _ = require('lodash');
 
-const { underscoreKeys } = require('../helpers/objectUtils');
+const { underscoreKeys } = require('../helpers/object_utils');
 
 exports.showSerializer = user => underscoreKeys(_.pick(user, 'id', 'firstName', 'lastName', 'email'));
